Fix some tests

+++ modules/task_2/shagov_m_fox/fox.h
@@ -17,5 +17,5 @@ bool isEqualMatrix(Matrix A, Matrix B);
 Matrix createRandomMatrix(size_t size);
 Matrix sequentialMatrixMultiplication(std::vector<double> A, std::vector<double> B, size_t BlockSize);
 Matrix sequentialBlockMatrixMultiplication(std::vector<double> A, std::vector<double> B, size_t Size);
-Matrix parallelBlockMatrixMultiplication(std::vector<double> A, std::vector<double> B, int Size);
+Matrix parallelBlockMatrixMultiplication(std::vector<double> A, std::vector<double> B, size_t Size);
 #endif  // MODULES_TASK_2_SHAGOV_M_FOX_FOX_H_
